tbb 29

+++ tasks/artyushkina_markirovka/tests/performance/main.cpp
@@ -2,8 +2,6 @@
 
 #include <cstddef>
 #include <cstdint>
-#include <iostream>
-#include <utility>
 
 #include "artyushkina_markirovka/common/include/common.hpp"
 #include "artyushkina_markirovka/seq/include/ops_seq.hpp"
@@ -39,7 +37,7 @@ class ArtyushkinaMarkirovkaPerfTestsBase : public ppc::util::BaseRunPerfTests<In
     int rows = static_cast<int>(output_data[0]);
     int cols = static_cast<int>(output_data[1]);
 
-    if (static_cast<int>(input_data_[0]) != rows || static_cast<int>(input_data_[1]) != cols) {
+    if (input_data_[0] != rows || input_data_[1] != cols) {
       return false;
     }
